backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sentence-transformers==2.2.2
transformers==4.29.2
huggingface-hub==0.25.2
torch>=2.0.0
faiss-cpu==1.7.4
pydantic==1.10.7
python-multipart==0.0.6
aiofiles==0.8.0
gunicorn==20.1.0
numpy<2.0.0
rank_bm25==0.2.2
nltk==3.9.1
requests==2.31.0
tqdm==4.67.1